# Use the new hashtable for core_namemap