--- a/test/lhash_test.c
+++ b/test/lhash_test.c
@@ -230,6 +230,7 @@ static int test_int_hashtable(void)
         NULL,
         NULL,
         NULL,
+        1,
         0,
     };
     INTKEY key;
@@ -407,6 +408,7 @@ static int test_hashtable_stress(void)
         NULL,              /* use default context */
         hashtable_intfree, /* our free function */
         hashtable_hash,    /* our hash function */
+        1,                 /* Check collisions */
         625000,            /* preset hash size */
     };
     HT *h;
@@ -625,6 +627,7 @@ static int test_hashtable_multithread(void)
         NULL,              /* use default context */
         hashtable_mt_free, /* our free function */
         NULL,              /* default hash function */
+        1,                 /* Check collisions */
         0,                 /* default hash size */
     };
     int ret = 0;