… crypto/cpt_err.c
… crypto/err/openssl.txt
… include/crypto/cryptoerr.h
… include/openssl/cryptoerr.h
… test/namemap_internal_test.c