This replaces LHASH in core_namemap with the new hashtable and adds the reverse mapping.

The new hashtable is used in lockless-read mode.

This is a draft PR as it is not really obvious this is performing better than the existing implementation.

Here are some perf test results as run on my laptop:

```
OLD:

for i in 1 10 100 500 1000 ; do ./evp_fetch $i ; done
Average time per fetch call: 0.118657us
Average time per fetch call: 0.937516us
Average time per fetch call: 5.082460us
Average time per fetch call: 23.874470us
Average time per fetch call: 31.097330us
[mraz@leccinum perf]$ for i in 1 10 100 500 1000 ; do ./evp_fetch $i ; done
Average time per fetch call: 0.121179us
Average time per fetch call: 0.825382us
Average time per fetch call: 5.013673us
Average time per fetch call: 22.839906us
Average time per fetch call: 45.688984us
[mraz@leccinum perf]$ for i in 1 10 100 500 1000 ; do ./evp_fetch $i ; done
Average time per fetch call: 0.120794us
Average time per fetch call: 0.817245us
Average time per fetch call: 5.227036us
Average time per fetch call: 23.454065us
Average time per fetch call: 35.714118us

NEW:

for i in 1 10 100 500 1000 ; do ./evp_fetch $i ; done
Average time per fetch call: 0.132573us
Average time per fetch call: 0.712213us
Average time per fetch call: 4.553993us
Average time per fetch call: 20.472451us
Average time per fetch call: 11.578301us
[mraz@leccinum perf]$ for i in 1 10 100 500 1000 ; do ./evp_fetch $i ; done
Average time per fetch call: 0.134698us
Average time per fetch call: 0.753527us
Average time per fetch call: 4.910886us
Average time per fetch call: 25.267473us
Average time per fetch call: 17.515930us
[mraz@leccinum perf]$ for i in 1 10 100 500 1000 ; do ./evp_fetch $i ; done
Average time per fetch call: 0.137634us
Average time per fetch call: 0.702508us
Average time per fetch call: 4.529108us
Average time per fetch call: 21.943141us
Average time per fetch call: 20.204932us


OLD:

$ for i in 1 10 100 500 1000 ; do ./pkeyread -f pem -k x25519 -t $i; done
[x25519 pem] 42.210110us
[x25519 pem] 81.899330us
[x25519 pem] 537.575520us
[x25519 pem] 1096.529970us
[x25519 pem] 699.809490us
[mraz@leccinum perf]$ for i in 1 10 100 500 1000 ; do ./pkeyread -f pem -k x25519 -t $i; done
[x25519 pem] 42.074580us
[x25519 pem] 81.642810us
[x25519 pem] 502.460940us
[x25519 pem] 1940.176050us
[x25519 pem] 815.381970us
[mraz@leccinum perf]$ for i in 1 10 100 500 1000 ; do ./pkeyread -f pem -k x25519 -t $i; done
[x25519 pem] 42.105730us
[x25519 pem] 77.209990us
[x25519 pem] 483.336400us
[x25519 pem] 1563.602100us
[x25519 pem] 1289.410440us

NEW:

$ for i in 1 10 100 500 1000 ; do ./pkeyread -f pem -k x25519 -t $i; done
[x25519 pem] 42.182840us
[x25519 pem] 82.937740us
[x25519 pem] 501.088150us
[x25519 pem] 2088.188980us
[x25519 pem] 725.168830us
[mraz@leccinum perf]$ for i in 1 10 100 500 1000 ; do ./pkeyread -f pem -k x25519 -t $i; done
[x25519 pem] 42.499780us
[x25519 pem] 81.852910us
[x25519 pem] 521.454130us
[x25519 pem] 796.237930us
[x25519 pem] 1504.101260us
[mraz@leccinum perf]$ for i in 1 10 100 500 1000 ; do ./pkeyread -f pem -k x25519 -t $i; done
[x25519 pem] 41.920050us
[x25519 pem] 86.067110us
[x25519 pem] 526.918730us
[x25519 pem] 954.933760us
[x25519 pem] 693.064710us


OLD:

$ for i in 1 10 100 500 1000 ; do ./handshake ~/osrc/openssl/test/certs $i ; done
Average time per handshake: 1032.356000us
Handshakes per second: 968.637738
Average time per handshake: 1908.849600us
Handshakes per second: 4722.084101
Average time per handshake: 15410.307100us
Handshakes per second: 5721.606101
Average time per handshake: 61155.006200us
Handshakes per second: 5639.098804
Average time per handshake: 93759.851700us
Handshakes per second: 5617.762917

NEW:

$ for i in 1 10 100 500 1000 ; do ./handshake ~/osrc/openssl/test/certs $i ; done
Average time per handshake: 1033.632800us
Handshakes per second: 967.437692
Average time per handshake: 1981.853100us
Handshakes per second: 4666.937038
Average time per handshake: 15161.255300us
Handshakes per second: 5824.233939
Average time per handshake: 61785.274700us
Handshakes per second: 5685.934023
Average time per handshake: 82916.679700us
Handshakes per second: 5630.820861
```

Unfortunately the values are unstable as running perf tests on laptop which is not 100% idle won't produce them.

I would summarize the results as that in the non-concurrent case the old code performs a tiny bit better. In concurrent case especially with high number of threads the new code performs better in the evp_fetch test. The handshake and pkeyread tests are imo inconclusive.

Before just dropping this PR I would like to measure the performance on some other HW first.