@t8m 2024-05-27
@nhorman with 1024 neighborhoods the collision limit was hit. I used 4096 as there will be external names added with 3rd party providers. This makes for 256KB hashtable.