@t8m 2024-05-27
Hmm... actually running the evp_fetch test directly on the work machine gives some better looking numbers:

Old code:

```
tomas@work:~/tools/perf$ for i in 1 10 50 100 500 1000 ; do echo -n Threads $i:' ' ; ./evp_fetch $i ; done
Threads 1: Average time per fetch call: 0.174428us
Threads 10: Average time per fetch call: 2.370862us
Threads 50: Average time per fetch call: 10.178295us
Threads 100: Average time per fetch call: 23.831959us
Threads 500: Average time per fetch call: 85.390427us
Threads 1000: Average time per fetch call: 184.885454us
tomas@work:~/tools/perf$ for i in 1 10 50 100 500 1000 ; do echo -n Threads $i:' ' ; ./evp_fetch $i ; done
Threads 1: Average time per fetch call: 0.173596us
Threads 10: Average time per fetch call: 2.143247us
Threads 50: Average time per fetch call: 10.857971us
Threads 100: Average time per fetch call: 25.316547us
Threads 500: Average time per fetch call: 98.130469us
Threads 1000: Average time per fetch call: 162.782784us
tomas@work:~/tools/perf$ for i in 1 10 50 100 500 1000 ; do echo -n Threads $i:' ' ; ./evp_fetch $i ; done
Threads 1: Average time per fetch call: 0.177012us
Threads 10: Average time per fetch call: 2.306977us
Threads 50: Average time per fetch call: 10.794114us
Threads 100: Average time per fetch call: 24.252512us
Threads 500: Average time per fetch call: 99.927294us
Threads 1000: Average time per fetch call: 152.574592us
tomas@work:~/tools/perf$ for i in 1 10 50 100 500 1000 ; do echo -n Threads $i:' ' ; ./evp_fetch $i ; done
Threads 1: Average time per fetch call: 0.173985us
Threads 10: Average time per fetch call: 2.303798us
Threads 50: Average time per fetch call: 11.181545us
Threads 100: Average time per fetch call: 24.965385us
Threads 500: Average time per fetch call: 69.697347us
Threads 1000: Average time per fetch call: 119.603823us
```

New code:
```
tomas@work:~/tools/perf$ for i in 1 10 50 100 500 1000 ; do echo -n Threads $i:' ' ; ./evp_fetch $i ; done
Threads 1: Average time per fetch call: 0.176699us
Threads 10: Average time per fetch call: 2.006918us
Threads 50: Average time per fetch call: 7.644519us
Threads 100: Average time per fetch call: 17.333350us
Threads 500: Average time per fetch call: 37.251312us
Threads 1000: Average time per fetch call: 50.760661us
tomas@work:~/tools/perf$ for i in 1 10 50 100 500 1000 ; do echo -n Threads $i:' ' ; ./evp_fetch $i ; done
Threads 1: Average time per fetch call: 0.176168us
Threads 10: Average time per fetch call: 2.024267us
Threads 50: Average time per fetch call: 8.956190us
Threads 100: Average time per fetch call: 16.553451us
Threads 500: Average time per fetch call: 47.369322us
Threads 1000: Average time per fetch call: 36.155592us
tomas@work:~/tools/perf$ for i in 1 10 50 100 500 1000 ; do echo -n Threads $i:' ' ; ./evp_fetch $i ; done
Threads 1: Average time per fetch call: 0.176002us
Threads 10: Average time per fetch call: 2.041097us
Threads 50: Average time per fetch call: 7.169573us
Threads 100: Average time per fetch call: 15.368748us
Threads 500: Average time per fetch call: 42.292506us
Threads 1000: Average time per fetch call: 48.808316us
tomas@work:~/tools/perf$ for i in 1 10 50 100 500 1000 ; do echo -n Threads $i:' ' ; ./evp_fetch $i ; done
Threads 1: Average time per fetch call: 0.176003us
Threads 10: Average time per fetch call: 1.879018us
Threads 50: Average time per fetch call: 7.290042us
Threads 100: Average time per fetch call: 18.238024us
Threads 500: Average time per fetch call: 40.959293us
Threads 1000: Average time per fetch call: 44.620931us
```

So this is consistently better with higher thread counts and not really worse otherwise.